controller

+++ src/main/java/com/nguyenanhtuyen/repository/PostRepository.java
@@ -9,19 +9,19 @@ import org.springframework.data.repository.query.Param;
 
 import com.nguyenanhtuyen.model.Post;
 
-public interface PostRepository extends JpaRepository<Post, Integer>{
+public interface PostRepository extends JpaRepository<Post, Long>{
 
 	@Query("SELECT p FROM Post p order by p.postedDate DESC")
 	public List<Post> findAll();
-	
+
 	@Query("SELECT p FROM Post p WHERE p.username=:username order by p.postedDate DESC")
 	public List<Post> findPostByUsername(@Param("username") String username);
-	
+
 	@Query("SELECT p FROM Post p WHERE p.id=:x")
-	public Post findPostById(@Param("x") Integer id);
+	public Post findPostById(@Param("x") Long id);
 
 	@Modifying
 	@Query("DELETE Post WHERE id=:x")
-	public void deletePostById(@Param("x") Integer id);
+	public void deletePostById(@Param("x") Long id);
 	
 }
